Before cleanup

+++ include/process.h
@@ -18,7 +18,7 @@ class Process {
   int pid = 0;
   std::string command = "none";
   //long int up_time = 0;
-  double cpu_usage = 0;
+  float cpu_usage = 0;
   //std::string ram = "";
   std::string user = "";
